+++ Changelog.md
@@ -2,6 +2,7 @@ HibernationFixup Changelog
 ============================
 #### v1.4.6
 - Added constants for macOS 13 support
+- Decouple WhenBatteryIsNotCharging/WhenBatteryIsAtWarnLevel and minimal capacity in parameter `hbfx-ahbm`
 
 #### v1.4.5
 - When battery level is critical, try to put macOS into sleep/hibernate mode only once per minute.